user-service/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
requests==2.31.0
prometheus-flask-exporter==0.23.0
Werkzeug==2.3.7
PyJWT==2.8.0
